Fix: Types

+++ Models/Barcode.cs
@@ -5,10 +5,10 @@ namespace api.Models
     public class Barcode 
     {
         public int Id { get; set; }
-        public int UPC { get; set; }
-        public int EAN { get; set; }
-        public int ISBN { get; set; }
-        public int JAN { get; set; }
-        public int ITF_14 { get; set; }
+        public long UPC { get; set; }
+        public long EAN { get; set; }
+        public long ISBN { get; set; }
+        public long JAN { get; set; }
+        public long ITF_14 { get; set; }
     }
 }
